pyproject.toml
[tool.poetry]
name = "fastapi_docker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79"
SQLAlchemy = "^1.4.39"
alembic = "^1.8.0"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
loguru = "^0.6.0"
typeguard = "^2.13.3"
Faker = "^13.15.0"
email-validator = "^1.2.1"
requests = "^2.28.1"
typer = "^0.6.1"
disposable-email-domains = "^0.0.79"
passlib = {extras = ["argon2"], version = "^1.7.4"}
pytz = "^2022.1"
langcodes = "^3.3.0"
python-stdnum = "^1.17"
flake8 = "^4.0.1"
Unidecode = "^1.3.4"
pendulum = "^2.1.2"
sentry-sdk = "^1.8.0"
user-agents = "^2.2.0"
boto3 = "^1.24.37"
python-multipart = "^0.0.5"
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.9.3"}
pydantic-factories = "^1.4.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
pytest-coverage = "^0.0"
isort = "^5.10.1"
autoflake = "^1.4"
sqlacodegen = "3.0.0rc1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
parallel = true
source = [
    "app",
    "tests"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    'if __name__ == "__main__":',
    "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
known_third_party = ["fastapi"]